requirements.txt
praw==7.7.1
python-dotenv==1.0.0
airtable-python-wrapper==0.15.3
requests==2.31.0
openai 